requirements.txt
flask
flask-sqlalchemy
pymysql
python-dotenv
cryptography